+++ Nyx/Source/Network/TCP/TCPServerSocket.cpp
@@ -40,7 +40,7 @@ namespace Nyx {
 
 		// ソケットの設定
 		srcAddr.sin_family = AF_INET;
-		srcAddr.sin_port = htons(port);
+		srcAddr.sin_port = htons(static_cast<ushort>(port));
 		srcAddr.sin_addr.S_un.S_addr = INADDR_ANY;
 		if(bind(srcSock, (struct sockaddr *)&srcAddr, sizeof(srcAddr)) != 0){
 			exit(-1);
